Removing fill attribute from ScrollTimeline
A test has been added to verify that startScrollOffset and
endScrollOffset still affect animations as they should.

Current implementation returns 0 for when the timeline
current time is before startScrollOffset and time_range for when
timeline current time is at, or after endScrollOffset. All based on the
spec:

https://drafts.csswg.org/scroll-animations-1/#current-time-algorithm

Removal of CC ScrollTimeline fill is included here because there is
conversion from main thread ScrollTimelines to CC ScrollTimelines in
scroll_timeline_util.cc and passing in a default fill value seems wrong.
CC ScrollTimelines currently do not support phase though, so the case
of current_time 0 will be ambiguous until that is implemented.

Bug: 1041621
Change-Id: I938679f0c61b42b588933bf9e5a8e7bd2776f529